+++ .storybook/preview.tsx
@@ -1,11 +1,13 @@
 import React from 'react'
-import { initialize, mswDecorator, mswLoader } from 'msw-storybook-addon'
+import { initialize, mswLoader } from 'msw-storybook-addon'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import type { Preview } from '@storybook/react'
 import '../src/app/styles/globals.css'
 import { themes } from '@storybook/theming'
 
-initialize()
+initialize({
+  onUnhandledRequest: 'bypass',
+})
 const queryClient = new QueryClient()
 
 const preview: Preview = {
